backend/requirements.txt
Django>=4.0
djangorestframework>=3.14.0
firebase-admin>=6.0.0